re-organize some functions and add tests
